refactor: prepare for gui feature
